Project/requirements.txt
click==6.7
cycler==0.10.0
Flask==0.12.2
imutils==0.4.5
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
matplotlib==2.1.1
numpy==1.13.3
olefile==0.44
opencv-contrib-python==3.3.1.11
Pillow==4.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
Werkzeug==0.13
